pyproject.toml
[tool.poetry]
name = "codereviewai"
version = "0.1.0"
description = ""
authors = ["DarthVaider <lyy1707@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
httpx = "^0.28.1"
pydantic = "^2.10.3"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.1"
setuptools = "^75.6.0"
redis = "^5.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
httpx = "^0.28.1"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.24.0"
fakeredis = "^2.26.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
